mcp-tools/web-automation/pyproject.toml
[project]
name = "web-automation"
version = "0.1.0"
description = "MCP server for web-based automation and migration tasks"
requires-python = ">=3.11"
dependencies = [
    "mcp>=0.1.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "duckdb>=0.9.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "aiofiles>=23.0.0",
    "tenacity>=8.2.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
web-automation = "web_automation.server:main"